# backend/src/api/openapi.yaml
openapi: 3.1.0
info:
  title: SkyCrop API
  version: 0.1.0
  description: |
    SkyCrop - Satellite-Based Paddy Field Management & Monitoring System
    MVP Auth and Health endpoints.
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.skycrop.com
    description: Production server
...
paths:
...
  /api/v1/satellite/tiles/{z}/{x}/{y}:
    get:
      tags: [Satellite]
      summary: Get Sentinel-2 tile via proxy
      description: Returns a tile image. Supports ETag and Cache-Control headers; Redis-backed caching.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: z
          required: true
          schema: { type: integer, minimum: 0 }
        - in: path
          name: x
          required: true
          schema: { type: integer, minimum: 0 }
        - in: path
          name: y
          required: true
          schema: { type: integer, minimum: 0 }
        - in: query
          name: bands
          schema:
            type: string
            description: CSV list of bands
            example: RGB,NIR
        - in: query
          name: date
          schema: { type: string, format: date }
        - in: query
          name: cloud_lt
          schema: { type: integer, minimum: 0, maximum: 100, default: 20 }
      responses:
        '200':
          description: Tile image
          headers:
            ETag:
              schema: { type: string }
            Cache-Control:
              schema: { type: string }
              example: public, max-age=21600
          content:
            image/png:
              schema: { type: string, format: binary }
            image/jpeg:
              schema: { type: string, format: binary }
        '304':
          description: Not Modified
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Upstream unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/satellite/preprocess:
    post:
      tags: [Satellite]
      summary: Submit preprocessing job (idempotent)
      description: Submits a preprocess job for bbox/date/bands/cloud-mask. Uses Idempotency-Key header.
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema: { type: string, minLength: 1, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bbox, date, bands]
              properties:
                bbox:
                  type: array
                  items: { type: number }
                  minItems: 4
                  maxItems: 4
                date:
                  type: string
                  format: date
                bands:
                  type: array
                  items: { type: string, enum: [RGB, NIR, SWIR, RED, GREEN, BLUE] }
                  minItems: 1
                cloud_mask:
                  type: boolean
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      job_id: { type: string }
                      status: { type: string, example: queued }
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/fields/{fieldId}/yield/predict:
    post:
      tags: [Yield Prediction]
      summary: Generate yield prediction for a field
      description: |
        Generates ML-powered yield prediction using Random Forest model.
        Analyzes NDVI history, weather data, and field characteristics.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: fieldId
          required: true
          schema:
            type: string
            format: uuid
          description: Field UUID
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                planting_date:
                  type: string
                  format: date
                  description: Planting date (defaults to current date)
                  example: '2024-01-15'
                crop_variety:
                  type: string
                  maxLength: 100
                  description: Rice variety (e.g., BG 300, BG 360)
                  example: 'BG 300'
                soil_type:
                  type: string
                  maxLength: 50
                  description: Soil type
                  example: 'clay-loam'
                price_per_kg:
                  type: number
                  format: float
                  description: Expected price per kg in LKR (defaults to 80)
                  example: 80
                model_version:
                  type: string
                  maxLength: 20
                  description: ML model version (defaults to latest)
                  example: '1.0.0'
      responses:
        '200':
          description: Prediction generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/YieldPredictionResult'
                  meta:
                    type: object
                    properties:
                      correlation_id: { type: string, format: uuid }
                      latency_ms: { type: integer }
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Field not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: ML service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/fields/{fieldId}/yield/predictions:
    get:
      tags: [Yield Prediction]
      summary: Get yield predictions for a field
      description: Retrieves historical yield predictions for a specific field
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: fieldId
          required: true
          schema:
            type: string
            format: uuid
          description: Field UUID
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Maximum number of predictions to return
        - in: query
          name: sort
          schema:
            type: string
            enum: [prediction_date, predicted_yield_per_ha, created_at]
            default: prediction_date
          description: Sort field
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
      responses:
        '200':
          description: Predictions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/YieldPrediction'
                  meta:
                    type: object
                    properties:
                      correlation_id: { type: string, format: uuid }
                      latency_ms: { type: integer }
                      cache_hit: { type: boolean }
                      count: { type: integer }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Field not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Email is invalid
            details:
              type: object
              nullable: true
      required: [success, error]
...
    YieldPredictionResult:
      type: object
      properties:
        prediction_id:
          type: string
          format: uuid
          description: Unique prediction ID
        field_id:
          type: string
          format: uuid
          description: Field UUID
        field_name:
          type: string
          description: Field name
        field_area_ha:
          type: number
          format: float
          description: Field area in hectares
          example: 2.5
        prediction_date:
          type: string
          format: date
          description: Date of prediction
        predicted_yield_per_ha:
          type: number
          format: float
          description: Predicted yield per hectare (kg/ha)
          example: 4800
        predicted_total_yield:
          type: number
          format: float
          description: Total predicted yield (kg)
          example: 12000
        confidence_interval:
          type: object
          properties:
            lower:
              type: number
              format: float
              example: 4200
            upper:
              type: number
              format: float
              example: 5400
        expected_revenue:
          type: number
          format: float
          description: Expected revenue in LKR
          example: 960000
        harvest_date_estimate:
          type: string
          format: date
          description: Estimated harvest date
        model_version:
          type: string
          description: ML model version used
          example: '1.0.0'
        features_used:
          type: object
          description: Features used for prediction
          properties:
            ndvi_avg: { type: number }
            ndvi_max: { type: number }
            ndvi_min: { type: number }
            ndvi_std: { type: number }
            rainfall_mm: { type: number }
            temp_avg: { type: number }
            humidity: { type: number }
            area_ha: { type: number }
        ml_response:
          type: string
          enum: [fresh, cached]
          description: Whether prediction came from ML service or cache
    YieldPrediction:
      type: object
      properties:
        prediction_id:
          type: string
          format: uuid
        field_id:
          type: string
          format: uuid
        prediction_date:
          type: string
          format: date
        predicted_yield_per_ha:
          type: number
          format: float
          example: 4800
        predicted_total_yield:
          type: number
          format: float
          example: 12000
        confidence_interval:
          type: object
          properties:
            lower: { type: number, format: float }
            upper: { type: number, format: float }
        expected_revenue:
          type: number
          format: float
          example: 960000
        harvest_date_estimate:
          type: string
          format: date
        model_version:
          type: string
          example: '1.0.0'
        actual_yield:
          type: number
          format: float
          nullable: true
          description: Actual yield if harvest completed
        accuracy_mape:
          type: number
          format: float
          nullable: true
          description: Mean Absolute Percentage Error if actual yield available
        created_at:
          type: string
          format: date-time
...