# backend/src/api/routes/satellite.routes.js
'use strict';

const express = require('express');
const Joi = require('joi');
const { authMiddleware } = require('../middleware/auth.middleware');
const { validateRequest, schemas } = require('../middleware/validation.middleware');
const SatelliteController = require('../controllers/satellite.controller');

const router = express.Router();

// Auth for all Satellite endpoints
router.use(authMiddleware);

// GET /api/v1/satellite/tiles/:z/:x/:y
router.get(
  '/tiles/:z/:x/:y',
  validateRequest(schemas.satelliteTileParams, 'params'),
  validateRequest(schemas.satelliteTileQuery, 'query'),
  SatelliteController.getTile
);

// Middleware: enforce request size cap (10 MB) for preprocess payloads
function enforceMaxBodySize(maxBytes) {
  return (req, res, next) => {
    try {
      const headerLen = req.headers['content-length']
        ? parseInt(req.headers['content-length'], 10)
        : null;
      if (Number.isFinite(headerLen) && headerLen > maxBytes) {
        return res.status(413)on({
          success: false,
          error: {
            code: 'PAYLOADTOOLARGE',
            message: `Request payload too large. Max ${maxBytes} bytes`,
          },
          meta: { timestamp: new Date().toISOString() },
        });
      }
      // Fallback: estimate from already-parsed body
      if (!headerLen && req.body && typeof req.body === 'object') {
        const approx = Buffer.byteLength(JSON.stringify(req.body));
        if (approx > maxBytes) {
          return res.status(413)on({
            success: false,
            error: {
              code: 'PAYLOADTOOLARGE',
              message: `Request payload too large. Max ${maxBytes} bytes`,
            },
            meta: { timestamp: new Date().toISOString() },
          });
        }
      }
      return next();
    } catch (e) {
      return next();
    }
  };
}

// POST /api/v1/satellite/preprocess
router.post(
  '/preprocess',
  enforceMaxBodySize(10 * 1024 * 1024), // 10 MB cap per integration plan
  validateRequest(schemas.satellitePreprocess, 'body'),
  SatelliteController.preprocess
);

// Optional: GET /api/v1/satellite/preprocess/:jobid (UUID v4 generated by crypto.randomUUID)
const jobIdParams = Joi.object({
  jobid: Joi.string()
    .guid({ version: ['uuidv4'] })
    .required(),
});
router.get(
  '/preprocess/:jobid',
  validateRequest(jobIdParams, 'params'),
  SatelliteController.getPreprocessStatus
);

module.exports = router;